+++ src/main/java/com/woowacourse/lotto/domain/PurchaseMoney.java
@@ -3,24 +3,24 @@ package com.woowacourse.lotto.domain;
 public class PurchaseMoney {
     private static final int MIN_PURCHASE_MONEY = 1000;
 
-    private final int money;
+    private final double money;
 
-    public PurchaseMoney(int money) {
+    public PurchaseMoney(double money) {
         checkMoney(money);
         this.money = money;
     }
 
-    private void checkMoney(int money) {
+    private void checkMoney(double money) {
         if (money < MIN_PURCHASE_MONEY) {
             throw new IllegalArgumentException("구입 금액이 부족합니다");
         }
     }
 
     public int getAvailableLottoSize() {
-        return money / MIN_PURCHASE_MONEY;
+        return (int) money / MIN_PURCHASE_MONEY;
     }
 
-    public int getProfitRatio(int lottoProfit) {
+    public double getProfitRatio(double lottoProfit) {
         return lottoProfit / money * 100;
     }
 
@@ -28,7 +28,7 @@ public class PurchaseMoney {
         return this.money >= money;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 }
